requirements.txt
nest-asyncio
pyngrok
ultralytics
opencv-contrib-python
google-cloud-vision
flask_cors